Split ktor-server-test-host module
