+++ gradle/compatibility.gradle
@@ -11,7 +11,8 @@ apiValidation {
         'ktor-client',
         'ktor-client-features',
         'ktor-client-benchmarks',
-        'ktor-server-benchmarks'
+        'ktor-server-benchmarks',
+        'ktor-server-test-host-suits',
     ].toSet()
 
     def projects = [].toSet()
