+++ ktor-server/ktor-server-cio/build.gradle
@@ -8,6 +8,7 @@ kotlin.sourceSets {
     }
     jvmTest.dependencies {
         api project(':ktor-server:ktor-server-test-host')
+        api project(':ktor-server:ktor-server-test-host-suits')
         api project(':ktor-server:ktor-server-core')
         api project(':ktor-client:ktor-client-cio')
     }
