+++ ktor-server/ktor-server-jetty/build.gradle
@@ -14,6 +14,7 @@ kotlin {
         jvmTest.dependencies {
             api project(':ktor-server:ktor-server-core')
             api project(':ktor-server:ktor-server-test-host')
+            api project(':ktor-server:ktor-server-test-host-suits')
             api group: 'org.eclipse.jetty', name: 'jetty-servlet', version: jetty_version
         }
     }
