+++ ktor-server/ktor-server-jetty/jvm/test/io/ktor/tests/server/jetty/JettyStressTest.kt
@@ -5,6 +5,6 @@
 package io.ktor.tests.server.jetty
 
 import io.ktor.server.jetty.*
-import io.ktor.server.testing.*
+import io.ktor.server.testing.suites.*
 
 class JettyStressTest : EngineStressSuite<JettyApplicationEngine, JettyApplicationEngineBase.Configuration>(Jetty)
