+++ ktor-server/ktor-server-jetty/ktor-server-jetty-test-http2/build.gradle
@@ -1,5 +1,6 @@
 kotlin.sourceSets.jvmTest.dependencies {
     api project(':ktor-server:ktor-server-test-host')
+    api project(':ktor-server:ktor-server-test-host-suits')
     api group: 'org.eclipse.jetty', name: 'jetty-servlet', version: jetty_version
     api project(':ktor-server:ktor-server-core')
     api project(":ktor-server:ktor-server-jetty")
