+++ ktor-server/ktor-server-test-host-suits/build.gradle
@@ -1,0 +1,11 @@
+description = ''
+
+kotlin.sourceSets.jvmMain.dependencies {
+    implementation project(':ktor-server:ktor-server-test-host')
+
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-debug:$coroutines_version"
+}
+
+dependencies {
+    jvmTestApi project(path: ':ktor-server:ktor-server-core', configuration: 'testOutput')
+}
