+++ ktor-server/ktor-server-test-host/api/ktor-server-test-host.api
@@ -1,17 +1,3 @@
-public abstract class io/ktor/server/testing/EngineStressSuite : io/ktor/server/testing/EngineTestBase {
-	public fun <init> (Lio/ktor/server/engine/ApplicationEngineFactory;)V
-	public fun getTimeout ()J
-	public final fun high load stress test ()V
-	public final fun multiple connections high pressure ()V
-	public final fun single connection high pressure ()V
-	public final fun single connection single thread no pipelining ()V
-	public final fun single connection single thread with pipelining ()V
-	public final fun test http upgrade ()V
-	public final fun test respond write ()V
-	public final fun test404 ()V
-	public final fun testLongResponse ()V
-}
-
 public abstract class io/ktor/server/testing/EngineTestBase : kotlinx/coroutines/CoroutineScope {
 	public static final field Companion Lio/ktor/server/testing/EngineTestBase$Companion;
 	public static field keyStore Ljava/security/KeyStore;
@@ -244,88 +230,5 @@ public final class io/ktor/server/testing/client/TestHttpClientEngine : io/ktor/
 
 public final class io/ktor/server/testing/client/TestHttpClientEngine$Companion : io/ktor/client/engine/HttpClientEngineFactory {
 	public fun create (Lkotlin/jvm/functions/Function1;)Lio/ktor/client/engine/HttpClientEngine;
-}
-
-public abstract class io/ktor/server/testing/suites/CompressionTestSuite : io/ktor/server/testing/EngineTestBase {
-	public fun <init> (Lio/ktor/server/engine/ApplicationEngineFactory;)V
-	public final fun testCompressionWriteToLarge ()V
-	public final fun testLocalFileContentRangeWithCompression ()V
-	public final fun testLocalFileContentWithCompression ()V
-	public final fun testStreamingContentWithCompression ()V
-}
-
-public abstract class io/ktor/server/testing/suites/ContentTestSuite : io/ktor/server/testing/EngineTestBase {
-	public static final field Companion Lio/ktor/server/testing/suites/ContentTestSuite$Companion;
-	public static final field classesDir Ljava/lang/String;
-	public static final field coreClassesDir Ljava/lang/String;
-	public fun <init> (Lio/ktor/server/engine/ApplicationEngineFactory;)V
-	public final fun testBinary ()V
-	public final fun testBinaryUsingChannel ()V
-	public fun testChunked ()V
-	public final fun testEchoBlocking ()V
-	public final fun testJarFileContent ()V
-	public final fun testLocalFileContent ()V
-	public final fun testLocalFileContentRange ()V
-	public final fun testMultipartFileUpload ()V
-	public final fun testMultipartFileUploadLarge ()V
-	public final fun testReceiveInputStream ()V
-	public final fun testRequestBodyAsyncEcho ()V
-	public final fun testRequestContentFormData ()V
-	public final fun testRequestContentInputStream ()V
-	public final fun testRequestContentString ()V
-	public final fun testSendTextWithContentType ()V
-	public final fun testStaticServe ()V
-	public final fun testStaticServeFromDir ()V
-	public final fun testStream ()V
-	public final fun testStreamNoFlush ()V
-	public final fun testTextContent ()V
-	public final fun testURIContent ()V
-	public final fun testURIContentLocalFile ()V
-}
-
-public final class io/ktor/server/testing/suites/ContentTestSuite$Companion {
-}
-
-public abstract class io/ktor/server/testing/suites/HttpServerTestSuite : io/ktor/server/testing/EngineTestBase {
-	public fun <init> (Lio/ktor/server/engine/ApplicationEngineFactory;)V
-	public final fun test404 ()V
-	public final fun testClosedConnection ()V
-	public final fun testConnectionReset ()V
-	public final fun testCookie ()V
-	public final fun testFormUrlEncoded ()V
-	public final fun testHeadRequest ()V
-	public final fun testHeader ()V
-	public final fun testHeadersReturnCorrectly ()V
-	public final fun testPathComponentsDecoding ()V
-	public final fun testProxyHeaders ()V
-	public final fun testRedirect ()V
-	public final fun testRedirectFromInterceptor ()V
-	public final fun testRemoteHost ()V
-	public final fun testRequestParameters ()V
-	public final fun testRequestParts ()V
-	public final fun testRequestTwiceInOneBufferWithKeepAlive ()V
-	public final fun testRequestTwiceNoKeepAlive ()V
-	public final fun testRequestTwiceWithKeepAlive ()V
-	public final fun testServerPush ()V
-	public final fun testStatusCodeDirect ()V
-	public final fun testStatusCodeViaResponseObject ()V
-	public final fun testStatusPages404 ()V
-	public fun testUpgrade ()V
-}
-
-public abstract class io/ktor/server/testing/suites/SustainabilityTestSuite : io/ktor/server/testing/EngineTestBase {
-	public fun <init> (Lio/ktor/server/engine/ApplicationEngineFactory;)V
-	public final fun testApplicationScopeCancellation ()V
-	public final fun testBigFile ()V
-	public final fun testBigFileHttpUrlConnection ()V
-	public final fun testBlockingConcurrency ()V
-	public fun testBlockingDeadlock ()V
-	public fun testChunkedWithVSpace ()V
-	public fun testChunkedWrongLength ()V
-	public final fun testEmbeddedServerCancellation ()V
-	public final fun testGetWithBody ()V
-	public final fun testIgnorePostContent ()V
-	public final fun testLoggerOnError ()V
-	public final fun testRepeatRequest ()V
 }
