+++ ktor-server/ktor-server-test-host/jvm/src/io/ktor/server/testing/Utils.kt
@@ -5,9 +5,6 @@
 package io.ktor.server.testing
 
 import io.ktor.http.*
-import java.io.*
-import java.util.zip.*
-import kotlin.test.*
 
 /**
  * [on] function receiver object
@@ -31,18 +28,6 @@ public fun on(comment: String, body: On.() -> Unit) = On.body()
 @Suppress("UNUSED_PARAMETER")
 public inline fun On.it(description: String, body: It.() -> Unit) = It.body()
 
-internal suspend fun assertFailsSuspend(block: suspend () -> Unit): Throwable {
-    var exception: Throwable? = null
-    try {
-        block()
-    } catch (cause: Throwable) {
-        exception = cause
-    }
-
-    assertNotNull(exception)
-    return exception
-}
-
 /**
  * Return parsed content type from the test response
  */
@@ -50,32 +35,3 @@ public fun TestApplicationResponse.contentType(): ContentType {
     val contentTypeHeader = requireNotNull(headers[HttpHeaders.ContentType])
     return ContentType.parse(contentTypeHeader)
 }
-
-internal fun InputStream.crcWithSize(): Pair<Long, Long> {
-    val checksum = CRC32()
-    val bytes = ByteArray(8192)
-    var count = 0L
-
-    do {
-        val rc = read(bytes)
-        if (rc == -1) {
-            break
-        }
-        count += rc
-        checksum.update(bytes, 0, rc)
-    } while (true)
-
-    return checksum.value to count
-}
-
-internal fun String.urlPath() = replace("\\", "/")
-
-internal class ExpectedException(message: String) : RuntimeException(message)
-
-internal fun loadTestFile(): File = listOf(
-    File("jvm/src"),
-    File("jvm/test"),
-    File("ktor-server/ktor-server-core/jvm/src")
-).filter { it.exists() }
-    .flatMap { it.walkBottomUp().asIterable() }
-    .first { it.extension == "kt" }
