+++ settings.gradle
@@ -21,6 +21,7 @@ include ':ktor-server:ktor-server-core'
 include ':ktor-server:ktor-server-tests'
 include ':ktor-server:ktor-server-host-common'
 include ':ktor-server:ktor-server-test-host'
+include ':ktor-server:ktor-server-test-host-suits'
 include ':ktor-server:ktor-server-jetty'
 include ':ktor-server:ktor-server-jetty:ktor-server-jetty-test-http2'
 include ':ktor-server:ktor-server-servlet'
